// src/app/shared/model/update-master-cust/update-cust-contact-info-obj.model.ts
export class UpdateCustContactInfoObj {
    CustCompanyContactPersonId: any;
    CustCompanyId: any;
    ContactPersonName: any;
    MrJobPositionCode: any;
    JobTitleName: any;
    MobilePhnNo1: any;
    MobilePhnNo2: any;
    Email1: any;
    Email2: any;
    MrGenderCode: any;
    CustAddrId: any;
    CustId: any;
    MrCustAddrTypeCode: any;
    Addr: any;
    AreaCode1: any;
    AreaCode2: any;
    AreaCode3: any;
    AreaCode4: any;
    City: any;
    Zipcode: any;
    RowVersionContactInfo: any;
    RowVersionContactInfoAddr: any;
}